requirements.txt
antiberty==0.1.3
biopython==1.78
numpy==1.26.4
pandas==2.2.2
torch==1.12.1
